requirements.txt
Flask>=2.0.3
Werkzeug>=2.0.3
bcrypt==3.2.0
PyJWT==2.1.0
cryptography==3.4.8
pydantic==1.8.2
pytest==6.2.5
pylint>=2.15.10
flask-limiter
python-dotenv==0.19.2
pydantic[email]==1.8.2
